+++ doc/source/examples/plot_ranges.py
@@ -16,7 +16,7 @@ Ranges of timeseries data
 # you can download the file from
 # https://github.com/IAMconsortium/pyam/tree/master/doc/source/tutorials.
 #
-# Make sure to place the file in the same folder as this script/notebook.
+# Make sure to place the data file in the same folder as this script/notebook.
 
 import matplotlib.pyplot as plt
 import pyam
@@ -33,8 +33,8 @@ df
 data = df.filter(scenario='CD-LINKS*',
                  variable='Emissions|CO2', region='World')
 
-fig, ax = plt.subplots()
-data.line_plot(ax=ax, color='scenario', fill_between=True)
+data.plot(color='scenario', fill_between=True)
+plt.tight_layout()
 plt.show()
 
 ###############################
@@ -45,8 +45,8 @@ plt.show()
 # or it can be provided specific arguments as a dictionary:
 # in this illustration, we choose a very low transparency value.
 
-fig, ax = plt.subplots()
-data.line_plot(ax=ax, color='scenario', fill_between=dict(alpha=0.15))
+data.plot(color='scenario', fill_between=dict(alpha=0.15))
+plt.tight_layout()
 plt.show()
 
 ###############################
@@ -57,7 +57,6 @@ plt.show()
 # range of data in the final time period using `final_ranges`. Similar to
 # `fill_between` it can either be true or have specific arguments.
 
-fig, ax = plt.subplots()
-data.line_plot(ax=ax, color='scenario', fill_between=True,
-               final_ranges=dict(linewidth=5))
+data.plot(color='scenario', fill_between=True, final_ranges=dict(linewidth=5))
+plt.tight_layout()
 plt.show()